Fixed click bug in Editor Topbar
Prior to this commit, clicking on any button under any entry in the topbar menu, mousedown events (but not mouseups) would not be caught by the topbar and would spread to whatever is below, performing an action. It can be observed easily in the script editor; open any script, then save with the mouse through the topbar menu entry; you will see it will make you select the text underneath the entry you clicked.